// src/pages/dashboard/gmb-local/types.ts
export interface CompanyData {
  nome_fantasia?: string | null;
  logradouro?: string | null;
  numero?: string | null;
  bairro?: string | null;
  cidade?: string | null;
  uf?: string | null;
  cep?: string | null;
  celular_atendimento?: string | null;
  site_oficial?: string | null;
  gmb_place_type?: string | null;
  gmb_place_type_secondary?: string | null;
  gmb_place_id?: string | null;
}